requirements.txt
pyinstaller
pystray
pillow==9.5
requests
pywin32
lxml
